refactor: Move block metadata and l1 origin structs into primitives
